# lib/api-zod/src/generated/types/chronotype.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SciVet API — citizen science papers, AI rigor checks, community reviews
 * OpenAPI spec version: 0.1.0
 */

export type Chronotype = (typeof Chronotype)[keyof typeof Chronotype];

export const Chronotype = {
  morning: "morning",
  evening: "evening",
  flexible: "flexible",
} as const;