# lib/api-zod/src/generated/types/focusCompleteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SciVet API — citizen science papers, AI rigor checks, community reviews
 * OpenAPI spec version: 0.1.0
 */

export interface FocusCompleteInput {
  /** @minimum 0 */
  focusedSeconds: number;
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  focusRating?: number | null;
  /**
   * @minimum 1
   * @maximum 5
   * @nullable
   */
  flowRating?: number | null;
  /** @nullable */
  reflection?: string | null;
}