# lib/api-zod/src/generated/types/focusGoalType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SciVet API — citizen science papers, AI rigor checks, community reviews
 * OpenAPI spec version: 0.1.0
 */

export type FocusGoalType = (typeof FocusGoalType)[keyof typeof FocusGoalType];

export const FocusGoalType = {
  read: "read",
  review: "review",
  revise: "revise",
  explore: "explore",
} as const;