# lib/api-zod/src/generated/types/focusHeartbeatInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SciVet API — citizen science papers, AI rigor checks, community reviews
 * OpenAPI spec version: 0.1.0
 */
import type { FocusSessionStatus } from "./focusSessionStatus";

export interface FocusHeartbeatInput {
  /** @minimum 0 */
  focusedSeconds: number;
  status?: FocusSessionStatus;
}