# lib/api-zod/src/generated/types/focusSessionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SciVet API — citizen science papers, AI rigor checks, community reviews
 * OpenAPI spec version: 0.1.0
 */
import type { FocusEvent } from "./focusEvent";
import type { FocusSession } from "./focusSession";

export interface FocusSessionDetail {
  session: FocusSession;
  events: FocusEvent[];
  qualifiesAsDeep: boolean;
}